requirements.txt
filetype
streamlit>=1.24.0
ollama-python
pydantic
pytest
pytest-asyncio
PyMuPDF
httpx 